mozilla-release/testing/web-platform/tests/fetch/sec-metadata/report.tentative.https.sub.html

<!DOCTYPE html>
<link rel="author" href="mtrzos@google.com" title="Maciek Trzos">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/sec-metadata/resources/helper.js></script>
<link id="style" href="https://foo.bar" rel="stylesheet">
<body></body>
<script>
<<<<<<< HEAD
  function generate_test(expected, name) {
    async_test(t => {
      t.step_timeout(_ => {
        return fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=report-" + name)
            .then(response => response.text())
            .then(text => assert_header_equals(text, expected))
            .then(_ => t.done());
      }, 1000);
    }, name + " report");
  }
||||||| merged common ancestors
  let counter = 0;
  document.addEventListener("securitypolicyviolation", (e) => {
    counter++;
    if (counter == 3) {
      promise_test(t => {
        expected = {"destination":"report", "site":"same-origin"};
        return fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=report-same-origin")
            .then(response => response.text())
            .then(text => assert_header_equals(text, expected));
      }, "Same-Origin report");
=======
  setup({ explicit_done: true });
  function generate_test(expected, name) {
    async_test(t => {
      t.step_timeout(_ => {
        return fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=report-" + name)
            .then(response => response.text())
            .then(text => assert_header_equals(text, expected))
            .then(_ => t.done());
      }, 1000);
    }, name + " report");
  }

  let counter = 0;
  document.addEventListener("securitypolicyviolation", (e) => {
    counter++;
    if (counter == 3) {
      generate_test({"dest":"report", "site":"same-origin", "user":"", "mode": "no-cors"}, "same-origin");
      generate_test({"dest":"report", "site":"same-site", "user":"", "mode": "no-cors"}, "same-site");
      generate_test({"dest":"report", "site":"cross-site", "user":"", "mode": "no-cors"}, "cross-site");
>>>>>>> upstream-releases

<<<<<<< HEAD
  test(_ => {
    let counter = 0;
    document.addEventListener("securitypolicyviolation", (e) => {
      counter++;
      if (counter == 3) {
        generate_test({"destination":"report", "site":"same-origin"}, "same-origin");
        generate_test({"destination":"report", "site":"same-site"}, "same-site");
        generate_test({"destination":"report", "site":"cross-site"}, "cross-site");
      }
    });
  }, "Initialization.");
||||||| merged common ancestors
      promise_test(t => {
        expected = {"destination":"report", "site":"same-site"};
        return fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=report-same-site")
            .then(response => response.text())
            .then(text => assert_header_equals(text, expected));
      }, "Same-site report");

      promise_test(t => {
        expected = {"destination":"report", "site":"cross-site"};
        return fetch("/fetch/sec-metadata/resources/record-header.py?retrieve=true&file=report-cross-site")
            .then(response => response.text())
            .then(text => assert_header_equals(text, expected));
      }, "Cross-site report");
    }
  });
=======
      done();
    }
  });
>>>>>>> upstream-releases
</script>
